+++ src/010-pipeline/build.ts
@@ -5,7 +5,7 @@ import { featureEnable } from "../000-config/features";
 import { ConfigProject } from "../000-config/parseConfig";
 import { CompilerContext } from "./context";
 import { funcCompile } from "../func/funcCompile";
-import { writeReport } from "../generator/writeReport";
+import { writeReport } from "../080-generator/writeReport";
 import { getRawAST } from "../050-grammar/store";
 import files from "../040-imports/stdlib";
 import { ILogger, Logger } from "./logger";
