+++ src/080-generator/Writer.ts
