+++ src/080-generator/builtins/AbiFunction.ts
@@ -1,8 +1,8 @@
-import { AstExpression } from "../050-grammar/ast";
-import { CompilerContext } from "../010-pipeline/context";
-import { WriterContext } from "../generator/Writer";
-import { TypeRef } from "../060-types/types";
-import { SrcInfo } from "../050-grammar";
+import { AstExpression } from "../../050-grammar/ast";
+import { CompilerContext } from "../../010-pipeline/context";
+import { WriterContext } from "../Writer";
+import { TypeRef } from "../../060-types/types";
+import { SrcInfo } from "../../050-grammar";
 
 export type AbiFunction = {
     name: string;
