+++ src/080-generator/emitter/createPadded.ts
