+++ src/080-generator/emitter/emit.ts
