+++ src/080-generator/writeProgram.ts
