+++ src/080-generator/writers/cast.ts
