+++ src/080-generator/writers/freshIdentifier.ts
