+++ src/080-generator/writers/id.ts
