+++ src/080-generator/writers/resolveFuncFlatPack.ts
