+++ src/080-generator/writers/resolveFuncFlatTypes.ts
