+++ src/080-generator/writers/resolveFuncPrimitive.ts
