+++ src/080-generator/writers/resolveFuncTupleType.ts
