+++ src/080-generator/writers/resolveFuncType.spec.ts
