+++ src/080-generator/writers/resolveFuncType.ts
