+++ src/080-generator/writers/writeConstant.ts
