+++ src/080-generator/writers/writeFunction.ts
