+++ src/080-generator/writers/writeInterfaces.ts
