+++ src/080-generator/writers/writeSerialization.spec.ts
